floating term: Use valid GTK class
Ghostty is picky about the window classes being a valid GTK application
ID. Switch to use a common class= argument for floating terminal
windows. This doesn't hurt alacritty, but will allow for ghostty to be
the default in the future.

This also fixes the audio menu not floating when launched from waybar if
alacritty isn't the default terminal.

Leave the existing types in hypr config for now in case they are being
used by user's configs.

Signed-off-by: Dylan Reid <dgreid@dylanreid.com>