screensaver: Use GTK compliant class
Similarly to floating windows, use a GTK ID compatible class argument so
ghostty doesn't reject it. Following changes will enable ghostty to be
used from more of these contexts.

This enables pkill and pgrep to use a smaller search term which will
make it easier to enable for kitty and ghostty.

Signed-off-by: Dylan Reid <dgreid@dylanreid.com>